playing with raycaster, trying to find height of ground
At current lat long (in index.html), we're colliding with things, but we're not yet sure which thing is the ground. Hypothesis: a mesh that contains a lot of points (faces) in its geometry (object.geometry.index.count) is likely to be the ground; others are likely to be sky or excessively coarse-grained meshes.